Handle equating of cells with the same name.
When cells with different names are equated, the equivalence of
cells with the same name is broken. This change allows the
equivalence of same named cells to be reestablished after it
is broken.

To equate A=A and A=B, use the following order
equate classes A B
equate classes A A